pyproject.toml
[tool.poetry]
name = "tg-assistant"
version = "0.1.0"
description = "Telegram bot for SUAI (aiogram + sqlite + chroma + ollama)"
authors = ["max <mahonovmax@gmail.com>"]
readme = "README.md"
packages = [{ include = "tg_assistant", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
aiogram = "^3.24.0"
pydantic-settings = "^2.0.0"
sqlalchemy = "^2.0.0"
aiosqlite = "^0.21.0"
alembic = "^1.13.0"
apscheduler = "^3.11.0"
chromadb = "^1.4.1"
aiohttp = "^3.10.0"
pypdf = "^5.0.0"
python-docx = "^1.1.0"
beautifulsoup4 = "^4.12.0"
[tool.poetry.group.dev.dependencies]
ruff = "^0.9.0"

[tool.poetry.scripts]
bot = "tg_assistant.main:run"

[build-system]
requires = ["poetry-core>=1.7.0"]
build-backend = "poetry.core.masonry.api"
